Add files via upload
Personal project of summarizing all of the money that I have made in the year that I have worked as a barista. I found which days I make the most money, what type of shift is the most lucrative, and how much I have made overall since earning the position.